Straightened features and dependencies, start on CI
