+++ examples/rt685s/Cargo.toml
@@ -16,6 +16,8 @@ repository = "https://github.com/OpenDevicePartnership/ec-slimloader"
 warnings = "deny"
 
 [workspace.dependencies]
+embassy-imxrt = { git = "https://github.com/Wassasin/embassy-imxrt.git", branch = "tmp/ec-slimloader-deps", default-features = false }
+
 partition-manager = { git = "https://github.com/OpenDevicePartnership/embedded-services.git", default-features = false }
 embassy-sync = "0.7.2"
