+++ examples/rt685s/application/memory.x
@@ -1,5 +1,4 @@
 MEMORY {
     FLASH    : ORIGIN = 0x00090000, LENGTH = 32K
-    /* FLASH    : ORIGIN = 0x08001000, LENGTH = 32K */
     RAM      : ORIGIN = 0x20098000, LENGTH = 1M   
 }
