+++ libs/.gitignore
@@ -1,2 +1,3 @@
 Cargo.lock
 /target
+/target_ci
