+++ libs/Cargo.toml
@@ -19,5 +19,5 @@ warnings = "deny"
 defmt = "0.3.7"
 defmt-or-log = { version = "0.2.2", default-features = false }
 embedded-storage-async = "0.4.1"
-
 embassy-sync = "0.7.2"
+log = "0.4"
